Move XLA Framework code from compiler/mlir to compiler/xla.
PiperOrigin-RevId: 487522997
