+++ tensorflow/compiler/xla/mlir/framework/ir/xla_framework.h
@@ -15,8 +15,8 @@ limitations under the License.
 
 // This file defines the operations and types used in the XLAFramework dialect.
 //
-#ifndef TENSORFLOW_COMPILER_MLIR_XLA_IR_XLA_FRAMEWORK_H_
-#define TENSORFLOW_COMPILER_MLIR_XLA_IR_XLA_FRAMEWORK_H_
+#ifndef TENSORFLOW_COMPILER_XLA_MLIR_FRAMEWORK_IR_XLA_FRAMEWORK_H_
+#define TENSORFLOW_COMPILER_XLA_MLIR_FRAMEWORK_IR_XLA_FRAMEWORK_H_
 
 #include "mlir/IR/Attributes.h"  // from @llvm-project
 #include "mlir/IR/BuiltinTypes.h"  // from @llvm-project
@@ -26,11 +26,11 @@ limitations under the License.
 #include "mlir/IR/OpImplementation.h"  // from @llvm-project
 
 #define GET_TYPEDEF_CLASSES
-#include "tensorflow/compiler/mlir/xla/ir/xla_framework_types.h.inc"
+#include "tensorflow/compiler/xla/mlir/framework/ir/xla_framework_types.h.inc"
 #define GET_OP_CLASSES
-#include "tensorflow/compiler/mlir/xla/ir/xla_framework.h.inc"
-#include "tensorflow/compiler/mlir/xla/ir/xla_framework_dialect.h.inc"
+#include "tensorflow/compiler/xla/mlir/framework/ir/xla_framework.h.inc"
+#include "tensorflow/compiler/xla/mlir/framework/ir/xla_framework_dialect.h.inc"
 
 #undef GET_OP_CLASSES
 
-#endif  // TENSORFLOW_COMPILER_MLIR_XLA_IR_XLA_FRAMEWORK_H_
+#endif  // TENSORFLOW_COMPILER_XLA_MLIR_FRAMEWORK_IR_XLA_FRAMEWORK_H_
