+++ tensorflow/compiler/xla/mlir/framework/tests/BUILD
@@ -1,0 +1,25 @@
+load("//tensorflow:tensorflow.default.bzl", "filegroup")
+load("//tensorflow/compiler/mlir:glob_lit_test.bzl", "glob_lit_tests")
+
+package(licenses = ["notice"])
+
+glob_lit_tests(
+    data = [":test_utilities"],
+    driver = "@llvm-project//mlir:run_lit.sh",
+    test_file_exts = [
+        "mlir",
+        "hlotxt",
+    ],
+)
+
+# Bundle together all of the test utilities that are used by tests.
+# This intentionally does not pull-in the top-level tf-opt to reduce the
+# dependencies.
+filegroup(
+    name = "test_utilities",
+    testonly = True,
+    data = [
+        "//tensorflow/compiler/xla/translate/mhlo_to_lhlo_with_xla:xla-translate-opt",
+        "@llvm-project//llvm:FileCheck",
+    ],
+)
